requirements.txt
gitpython>=3.1.31
python-dotenv>=1.0.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-openai>=0.0.2
faiss-cpu>=1.7.3
pandas>=2.2.3
matplotlib>=3.10.3
seaborn>=0.13.2
streamlit>=1.45.1
watchdog>=6.0.0